+++ books_resale/package.json
@@ -16,7 +16,7 @@
     "react-dom": "^18.2.0",
     "react-file-base64": "^1.0.3",
     "react-redux": "^8.0.5",
-    "react-router-dom": "^6.8.1",
+    "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.1",
     "redux-thunk": "^2.4.2",
