+++ books_resale/src/components/Form/form.css
@@ -1,4 +1,4 @@
 .formPaper{
-  background: #faf8f5;
-  /* background-image: url('../../images/Background.png'); */
+  /* background: #faf8f5; */
+  background-image: url('file:///C:/Users/dell/Downloads/bg_image.png');
 }
