+++ books_resale/src/index.css
@@ -5,8 +5,10 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background: #faf8f5;
-  /* background-image: url('./images/Background.png'); */
+  /* background: #faf8f5; */
+  background-image: url('file:///C:/Users/dell/Downloads/website_bg%20(1).png');
+  background-size:100%;
+  /* background-repeat: no-repeat; */
 }
 
 code {
